Move data logic for schedule feature to commons module
