+++ commons/src/main/java/com/droidcon/commons/data/schedule/disk/SessionsStorage.kt
@@ -1,7 +1,6 @@
-package com.droidcon.schedule.data.disk
+package com.droidcon.commons.data.schedule.disk
 
-import com.droidcon.commons.data.db.SessionsDao
-import com.droidcon.schedule.data.SessionData
+import com.droidcon.commons.data.schedule.SessionData
 import javax.inject.Inject
 import javax.inject.Singleton
