+++ commons/src/main/java/com/droidcon/commons/data/schedule/network/SessionsDTOs.kt
@@ -1,6 +1,9 @@
-package com.droidcon.schedule.data
+package com.droidcon.commons.data.schedule.network
 
-data class SessionData(
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SessionDto(
     val id: String,
     val title: String,
     val description: String?,
@@ -8,6 +11,6 @@ data class SessionData(
     val endsAt: String,
     val isServiceSession: Boolean,
     val isPlenumSession: Boolean,
-    val speakers: List<String>,
+    val speakers: List<String> = emptyList(),
     val roomId: Int?
 )
