+++ schedule/src/main/java/com/droidcon/schedule/domain/SessionsUseCases.kt
@@ -1,6 +1,6 @@
 package com.droidcon.schedule.domain
 
-import com.droidcon.schedule.data.SessionsRepository
+import com.droidcon.commons.data.schedule.SessionsRepository
 import java.util.*
 import javax.inject.Inject
